Added a Bash script called eval-script. This script will create a file called cells, fill it with a string consisting of '0' 49 times, '1', and '0' 49 more times. This file is then used to evaluate the cellular automata rules, starting from 1 through 256, and storing the results in a file called 'evals'. Also, renamed util.c to util.h and removed util.h from version control.
Added a function to  automata.c that will randomly apply a rule selected from a list of 'interesting' rules at each step of evolution. This method of evolution will be used if the executable is called with no parameters. To accomodate this, a function pointer was added to the code to allow a runtime decision of the evolution method to be performed. The executable take up from zero to two parameters: the first being the rule to be evaluated, the second the number of steps to evolve.

Finally, some rules may exhibit strange behavior at the boundaries of execution. This is because cells at the far left and right edges of the evaluation matrix are treated as if they have cells containing 0 to their left and right respectively. In some rules, this may result in strange behavior occuring at the edges of the view window (rule 73 is an example of this occurring). There is currently not a clean solution to this problem, as there does not seem to be a general, obvious manner to treat boundary cells in a way that does not perturb the normal evolution of the automata.